test: hygiene pass 2 — owners keep assertions, names match bodies
